Move FullscreenPagerAdapter to adapters package
